requirements.txt
line-bot-sdk
streamlit
openai